+++ src/common.rs
@@ -1,19 +1,29 @@
 //! Common functionality used by the mock implementations.
 
-use std::fmt::Debug;
-use std::sync::{Arc, Mutex};
+use std::{
+    collections::VecDeque,
+    fmt::Debug,
+    sync::{Arc, Mutex},
+    thread,
+};
 
-/// Generic Mock implementation
+/// Generic mock implementation.
 ///
-/// This supports the specification and evaluation of expectations to allow automated testing of hal drivers.
-/// Mismatches between expectations will cause runtime assertions to assist in locating the source of the fault.
+/// ⚠️ **Do not use this directly as end user! This is only a building block
+/// for creating mocks.**
+///
+/// This type supports the specification and evaluation of expectations to
+/// allow automated testing of hal drivers. Mismatches between expectations
+/// will cause runtime assertions to assist in locating the source of the
+/// fault.
 ///
 /// Note that the implementation uses an `Arc<Mutex<...>>` internally, so a
 /// cloned instance of the mock can be used to check the expectations of the
 /// original instance that has been moved into a driver.
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 pub struct Generic<T: Clone + Debug + PartialEq> {
-    expected: Arc<Mutex<(usize, Vec<T>)>>,
+    expected: Arc<Mutex<VecDeque<T>>>,
+    done_called: Arc<Mutex<DoneCallDetector>>,
 }
 
 impl<'a, T: 'a> Generic<T>
@@ -28,7 +38,8 @@ where
         E: IntoIterator<Item = &'a T>,
     {
         let mut g = Generic {
-            expected: Arc::new(Mutex::new((0, vec![]))),
+            expected: Arc::new(Mutex::new(VecDeque::new())),
+            done_called: Arc::new(Mutex::new(DoneCallDetector::new())),
         };
 
         g.expect(expected);
@@ -38,38 +49,26 @@ where
 
     /// Set expectations on the interface
     ///
-    /// This is a list of transactions to be executed in order
-    /// Note that setting this will overwrite any existing expectations
+    /// This is a list of transactions to be executed in order. Note that
+    /// setting this will overwrite any existing expectations.
     pub fn expect<E>(&mut self, expected: E)
     where
         E: IntoIterator<Item = &'a T>,
     {
-        let v: Vec<T> = expected.into_iter().cloned().collect();
-        let mut e = self.expected.lock().unwrap();
-        e.0 = 0;
-        e.1 = v;
+        let v: VecDeque<T> = expected.into_iter().cloned().collect();
+
+        let mut expected = self.expected.lock().unwrap();
+        let mut done_called = self.done_called.lock().unwrap();
+        *expected = v;
+        done_called.reset();
     }
 
-    /// Assert that all expectations on a given Mock have been met
+    /// Assert that all expectations on a given mock have been consumed.
     pub fn done(&mut self) {
-        let e = self.expected.lock().unwrap();
-        assert_eq!(
-            e.0,
-            e.1.len(),
-            "Mock call number(left) and expectations(right) do not match"
-        );
-    }
-}
+        self.done_called.lock().unwrap().mark_as_called();
 
-/// Clone allows a single mock to be duplicated for control and evaluation
-impl<T> Clone for Generic<T>
-where
-    T: Clone + Debug + PartialEq,
-{
-    fn clone(&self) -> Self {
-        Generic {
-            expected: self.expected.clone(),
-        }
+        let e = self.expected.lock().unwrap();
+        assert!(e.is_empty(), "Not all expectations consumed");
     }
 }
 
@@ -80,9 +79,60 @@ where
 {
     type Item = T;
     fn next(&mut self) -> Option<Self::Item> {
-        let mut e = self.expected.lock().unwrap();
-        e.0 += 1;
-        e.1.get(e.0 - 1).cloned()
+        self.expected.lock().unwrap().pop_front()
+    }
+}
+
+/// Struct used to detect whether or not the `.done()` method was called.
+#[derive(Debug)]
+pub(crate) struct DoneCallDetector {
+    called: bool,
+}
+
+impl DoneCallDetector {
+    pub(crate) fn new() -> Self {
+        Self { called: false }
+    }
+
+    /// Mark the `.done()` method as called.
+    ///
+    /// Note: When calling this method twice, an assertion failure will be
+    /// triggered.
+    pub(crate) fn mark_as_called(&mut self) {
+        assert!(!self.called, "The `.done()` method was called twice!");
+        self.called = true;
+    }
+
+    /// Reset the detector.
+    pub(crate) fn reset(&mut self) {
+        self.called = false;
+    }
+}
+
+impl Drop for DoneCallDetector {
+    fn drop(&mut self) {
+        // Ensure that the `.done()` method was called on the mock before
+        // dropping.
+        if !self.called && !thread::panicking() {
+            let msg = "WARNING: A mock (from embedded-hal-mock) was dropped \
+                       without calling the `.done()` method. \
+                       See https://github.com/dbrgn/embedded-hal-mock/issues/34 \
+                       for more details.";
+
+            // Note: We cannot use the print macros here, since they get
+            // captured by the Cargo test runner. Instead, write to stderr
+            // directly.
+            use std::io::Write;
+            let mut stderr = std::io::stderr();
+            stderr.write_all(b"\x1b[31m").ok();
+            stderr.write_all(msg.as_bytes()).ok();
+            stderr.write_all(b"\x1b[m\n").ok();
+            stderr.flush().ok();
+
+            // Panic. This probably results in an abort:
+            // https://doc.rust-lang.org/std/ops/trait.Drop.html#panics
+            panic!("{}", msg);
+        }
     }
 }
 
@@ -90,13 +140,41 @@ where
 mod tests {
     use super::*;
 
-    #[test]
-    fn test_generic_mock() {
-        let expectations = [0u8, 1u8];
-        let mut mock: Generic<u8> = Generic::new(&expectations);
+    mod generic_mock {
+        use super::*;
 
-        assert_eq!(mock.next(), Some(0u8));
-        assert_eq!(mock.next(), Some(1u8));
-        assert_eq!(mock.next(), None);
+        #[test]
+        fn success() {
+            let expectations = [0u8, 1u8];
+            let mut mock: Generic<u8> = Generic::new(&expectations);
+
+            assert_eq!(mock.next(), Some(0u8));
+            assert_eq!(mock.next(), Some(1u8));
+            assert_eq!(mock.next(), None);
+            assert_eq!(mock.next(), None);
+
+            mock.done();
+        }
+
+        #[test]
+        #[should_panic]
+        fn panic_if_drop_not_called() {
+            let expectations = [0u8, 1u8];
+            let mut mock: Generic<u8> = Generic::new(&expectations);
+            assert_eq!(mock.next(), Some(0u8));
+            assert_eq!(mock.next(), Some(1u8));
+            // Note: done() not called
+        }
+
+        #[test]
+        #[should_panic]
+        fn panic_if_drop_called_twice() {
+            let expectations = [0u8, 1u8];
+            let mut mock: Generic<u8> = Generic::new(&expectations);
+            assert_eq!(mock.next(), Some(0u8));
+            assert_eq!(mock.next(), Some(1u8));
+            mock.done();
+            mock.done(); // This will panic
+        }
     }
 }